# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Herpa Agentic GraphRAG Backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/files/upload": {
      "post": {
        "tags": [
          "Attachments"
        ],
        "summary": "Upload",
        "operationId": "upload_api_files_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_api_files_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/attachments/presign-upload": {
      "post": {
        "tags": [
          "Attachments"
        ],
        "summary": "Presign",
        "operationId": "presign_api_v1_attachments_presign_upload_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresignUploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresignUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/attachments/complete": {
      "post": {
        "tags": [
          "Attachments"
        ],
        "summary": "Complete",
        "operationId": "complete_api_v1_attachments_complete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttachmentCompleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttachmentStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/files/{attachment_id}/status": {
      "get": {
        "tags": [
          "Attachments"
        ],
        "summary": "Attachment Status",
        "operationId": "attachment_status_api_files__attachment_id__status_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "attachment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Attachment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttachmentStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/files/{attachment_id}/retry": {
      "post": {
        "tags": [
          "Attachments"
        ],
        "summary": "Retry",
        "operationId": "retry_api_files__attachment_id__retry_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "attachment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Attachment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttachmentStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/attachments/{attachment_id}": {
      "delete": {
        "tags": [
          "Attachments"
        ],
        "summary": "Delete",
        "operationId": "delete_api_v1_attachments__attachment_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "attachment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Attachment Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/analytics": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Analytics",
        "operationId": "analytics_api_admin_analytics_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminAnalytics"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/admin/users": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Users",
        "operationId": "users_api_admin_users_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/users/{user_id}": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "User Detail",
        "operationId": "user_detail_api_v1_admin_users__user_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/users/role": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Role",
        "operationId": "role_api_admin_users_role_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRoleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/users/{user_id}/status": {
      "patch": {
        "tags": [
          "Admin"
        ],
        "summary": "Update Status",
        "operationId": "update_status_api_v1_admin_users__user_id__status_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserStatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/usage/features": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Feature Usage",
        "operationId": "feature_usage_api_v1_admin_usage_features_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/usage/models": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Model Usage",
        "operationId": "model_usage_api_v1_admin_usage_models_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/usage/storage": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Storage Usage",
        "operationId": "storage_usage_api_v1_admin_usage_storage_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Storage Usage Api V1 Admin Usage Storage Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/audit-logs": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Audit Logs",
        "operationId": "audit_logs_api_v1_admin_audit_logs_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccountStatus": {
        "type": "string",
        "enum": [
          "active",
          "suspended",
          "deleted"
        ],
        "title": "AccountStatus"
      },
      "AdminAnalytics": {
        "properties": {
          "total_users": {
            "type": "integer",
            "title": "Total Users",
            "default": 0
          },
          "total_messages": {
            "type": "integer",
            "title": "Total Messages",
            "default": 0
          },
          "total_chats": {
            "type": "integer",
            "title": "Total Chats",
            "default": 0
          },
          "active_users_today": {
            "type": "integer",
            "title": "Active Users Today",
            "default": 0
          },
          "messages_today": {
            "type": "integer",
            "title": "Messages Today",
            "default": 0
          },
          "total_recommendations": {
            "type": "integer",
            "title": "Total Recommendations",
            "default": 0
          },
          "total_attachments": {
            "type": "integer",
            "title": "Total Attachments",
            "default": 0
          },
          "total_quiz_attempts": {
            "type": "integer",
            "title": "Total Quiz Attempts",
            "default": 0
          },
          "error_rate": {
            "type": "number",
            "title": "Error Rate",
            "default": 0
          },
          "average_latency_ms": {
            "type": "number",
            "title": "Average Latency Ms",
            "default": 0
          }
        },
        "type": "object",
        "title": "AdminAnalytics"
      },
      "ApplicationRole": {
        "type": "string",
        "enum": [
          "admin",
          "user"
        ],
        "title": "ApplicationRole"
      },
      "AttachmentCompleteRequest": {
        "properties": {
          "object_key": {
            "type": "string",
            "title": "Object Key"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          }
        },
        "type": "object",
        "required": [
          "object_key",
          "filename",
          "content_type",
          "size_bytes"
        ],
        "title": "AttachmentCompleteRequest"
      },
      "AttachmentInfo": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "stored_filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stored Filename"
          },
          "mime_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mime Type"
          },
          "preview_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preview Url"
          },
          "processing_status": {
            "type": "string",
            "enum": [
              "uploading",
              "queued",
              "processing",
              "completed",
              "failed"
            ],
            "title": "Processing Status",
            "default": "queued"
          },
          "detected_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detected Type"
          },
          "verification_status": {
            "type": "string",
            "title": "Verification Status",
            "default": "pending"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence",
            "default": 0.0
          }
        },
        "type": "object",
        "required": [
          "id",
          "filename"
        ],
        "title": "AttachmentInfo"
      },
      "AttachmentStatusResponse": {
        "properties": {
          "attachment_id": {
            "type": "string",
            "title": "Attachment Id"
          },
          "processing_status": {
            "type": "string",
            "enum": [
              "uploading",
              "queued",
              "processing",
              "completed",
              "failed"
            ],
            "title": "Processing Status"
          },
          "progress": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Progress"
          },
          "verification_status": {
            "type": "string",
            "title": "Verification Status"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "extracted_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extracted Text"
          },
          "detected_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detected Type"
          },
          "retryable": {
            "type": "boolean",
            "title": "Retryable",
            "default": false
          },
          "error": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "attachment_id",
          "processing_status",
          "progress",
          "verification_status",
          "confidence"
        ],
        "title": "AttachmentStatusResponse"
      },
...
      "Body_upload_api_files_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_api_files_upload_post"
      },
...
      "FileUploadResponse": {
        "properties": {
          "file_id": {
            "type": "string",
            "title": "File Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "extracted_text": {
            "type": "string",
            "title": "Extracted Text",
            "default": ""
          },
          "content_type": {
            "type": "string",
            "title": "Content Type"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "attachment": {
            "$ref": "#/components/schemas/AttachmentInfo"
          },
          "context": {
            "additionalProperties": true,
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "file_id",
          "filename",
          "content_type",
          "attachment"
        ],
        "title": "FileUploadResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PresignUploadRequest": {
        "properties": {
          "filename": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Filename"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type"
          },
          "size_bytes": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Size Bytes"
          },
          "purpose": {
            "type": "string",
            "enum": [
              "chat",
              "avatar"
            ],
            "title": "Purpose",
            "default": "chat"
          }
        },
        "type": "object",
        "required": [
          "filename",
          "content_type",
          "size_bytes"
        ],
        "title": "PresignUploadRequest"
      },
      "PresignUploadResponse": {
        "properties": {
          "attachment_id": {
            "type": "string",
            "title": "Attachment Id"
          },
          "object_key": {
            "type": "string",
            "title": "Object Key"
          },
          "upload_url": {
            "type": "string",
            "title": "Upload Url"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In"
          }
        },
        "type": "object",
        "required": [
          "attachment_id",
          "object_key",
          "upload_url",
          "expires_in"
        ],
        "title": "PresignUploadResponse"
      },
...
      "UpdateUserRoleRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "role": {
            "$ref": "#/components/schemas/ApplicationRole"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "role"
        ],
        "title": "UpdateUserRoleRequest"
      },
      "UpdateUserStatusRequest": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/AccountStatus"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "UpdateUserStatusRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}